requirements.txt
python-dotenv>=1.0.0
openai>=1.3.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
wikipedia>=1.4.0
requests>=2.31.0